Code refactor & license headers
