+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/identification/sirius/FingerIdWebMethodTask.java
@@ -1,17 +1,24 @@
-package net.sf.mzmine.modules.peaklistmethods.identification.sirius;/*
- * (C) Copyright 2015-2017 by MSDK Development Team
+/*
+ * Copyright 2006-2018 The MZmine 2 Development Team
  *
- * This software is dual-licensed under either
+ * This file is part of MZmine 2.
  *
- * (a) the terms of the GNU Lesser General Public License version 2.1 as published by the Free
- * Software Foundation
+ * MZmine 2 is free software; you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software
+ * Foundation; either version 2 of the License, or (at your option) any later
+ * version.
  *
- * or (per the licensee's choosing)
+ * MZmine 2 is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
  *
- * (b) the terms of the Eclipse Public License v1.0 as published by the Eclipse Foundation.
+ * You should have received a copy of the GNU General Public License along with
+ * MZmine 2; if not, write to the Free Software Foundation, Inc., 51 Franklin St,
+ * Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-import com.sun.xml.internal.bind.v2.runtime.unmarshaller.XsiNilLoader.Single;
+package net.sf.mzmine.modules.peaklistmethods.identification.sirius;
+
 import de.unijena.bioinf.ChemistryBase.ms.Ms2Experiment;
 import io.github.msdk.MSDKException;
 import io.github.msdk.datamodel.IonAnnotation;
@@ -19,7 +26,6 @@ import io.github.msdk.id.sirius.FingerIdWebMethod;
 import io.github.msdk.id.sirius.SiriusIonAnnotation;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
 import net.sf.mzmine.taskcontrol.AbstractTask;
 import net.sf.mzmine.taskcontrol.TaskStatus;
 import org.openscience.cdk.tools.manipulator.MolecularFormulaManipulator;
